Restore 100ms ghost cleanup via the shared daily stale-rooms cron
Vercel rejected a 5th cron entry, so instead of a dedicated /30-min schedule the
100ms ghost sweep now runs daily from the existing /api/cron/juke-stale-rooms
job. Extract the sweep into src/lib/social/sweep100msRooms.ts; the dedicated
/api/cron/100ms-stale-rooms route stays as a manual/on-demand trigger (not
scheduled). Real-time cleanup still comes from the 100ms webhook; this is the
backstop. No vercel.json cron change (stays at 4).
